requirements.txt
autopep8==1.5.7
dnspython==2.1.0
et-xmlfile==1.1.0
loguru==0.5.3
numpy==1.21.2
openpyxl==3.0.7
pandas==1.3.2
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
toml==0.10.2
